Moved some Prisma code to the database package
