+++ internal-packages/database/src/index.ts
@@ -1,1 +1,2 @@
 export * from "@prisma/client";
+export * from "./transaction";
